refactor(compiler): rename module
