+++ kaba_compiler/src/semantic/expression/index_access.rs
@@ -4,7 +4,7 @@ use crate::{
         error::Result,
         expression,
         state::AnalyzerState,
-        types::{assert, Type},
+        typ::{assert, Type},
     },
     AstNodeVariant,
 };
@@ -45,7 +45,7 @@ fn unwrap_index<'src, 'a>(node: &'a AstNode<'src>) -> &'a AstNode<'src> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::semantic::{test_util::assert_expr_type, types::IntType};
+    use crate::semantic::{test_util::assert_expr_type, typ::IntType};
 
     #[test]
     fn index_accessing() {
